requirements.txt
attrs==22.2.0
bcrypt==4.0.1
bitmask @ git+https://github.com/dogeystamp/bitmask@8524113fcdc22a570bda77d440374f5f269fdb79
black==23.3.0
click==8.1.3
coverage==7.2.1
exceptiongroup==1.1.0
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.3
Flask-Testing==0.8.1
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pluggy==1.0.0
PyJWT==2.6.0
pytest==7.2.2
pytest-cov==4.0.0
pytest-env==0.8.1
PyYAML==6.0
six==1.16.0
SQLAlchemy==2.0.5.post1
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.2.3
